Implement newtype wrapper around NonNull<Obj>
