+++ src/types/value.rs
@@ -1,0 +1,31 @@
+use std::fmt::Display;
+
+use super::obj_ref::ObjRef;
+
+#[derive(Debug, Clone, Copy, PartialEq)]
+pub enum Value {
+    Number(f64),
+    Boolean(bool),
+    Nil,
+    Obj(ObjRef),
+}
+
+impl Value {
+    pub fn is_falsey(&self) -> bool {
+        matches!(self, Value::Nil | Value::Boolean(false))
+    }
+}
+
+impl Display for Value {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            Value::Number(n) => write!(f, "{}", n),
+            Value::Boolean(b) => write!(f, "{}", b),
+            Value::Nil => write!(f, "Nil"),
+            Value::Obj(obj) => {
+                //SAFETY: managed by GC
+                unsafe { write!(f, "{}", obj) }
+            }
+        }
+    }
+}
